Replace interp2d with RectBivariateSpline
interp2d is gone in scipy 1.14, and RecBivariateSpline is 5x faster
to evaluate.

While the numerical results are identical, RectBivariateSpline
returns a 2D array for scalar inputs, whereas interp2d returned a
1D array.  Work around this by removing all 1-entry dimensions from
the result.  A notable exception is the acceptance in StandardLLH,
which appears to have relied on the result being 1D (i.e.  tests
fail if the interpolation returns a scalar).

Fixes #270